# docs/app/swagger.yaml
basePath: /
definitions:
  models.AppUser:
    properties:
      createdAt:
        type: string
      email:
        type: string
      firstName:
        description: Имя
        type: string
      id:
        type: integer
      isActive:
        description: Активен ли пользователь
        type: boolean
      lastName:
        description: Фамилия
        type: string
      middleName:
        description: Отчество (может быть пустым)
        type: string
      passwordHash:
        type: string
      role:
        type: string
      updatedAt:
        type: string
    type: object
...
  models.CreateUserRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      middle_name:
        type: string
      password:
        minLength: 6
        type: string
      role:
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - role
    type: object
...
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  models.LoginResponse:
    properties:
      expires_at:
        type: integer
      token:
        type: string
      user:
        $ref: '#/definitions/models.AppUser'
    type: object
...
info:
  contact: {}
  description: API for app endpoints with JWT authentication
  title: Vector App API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Вход в систему с email и паролем, возвращает JWT токен
      parameters:
      - description: Данные для входа
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешная аутентификация
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Неверные данные
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Неверный email или пароль
          schema:
            additionalProperties: true
            type: object
      summary: Аутентификация пользователя
      tags:
      - auth
  /auth/profile:
    get:
      consumes:
      - application/json
      description: Получить информацию о текущем аутентифицированном пользователе
      produces:
      - application/json
      responses:
        "200":
          description: Профиль пользователя
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Пользователь не аутентифицирован
          schema:
            additionalProperties: true
            type: object
      summary: Получить профиль текущего пользователя
      tags:
      - auth
...
  /auth/users:
    post:
      consumes:
      - application/json
      description: Создать нового пользователя в системе. Доступно только администраторам.
      parameters:
      - description: Данные пользователя
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь создан
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверные данные
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Недостаточно прав
          schema:
            additionalProperties: true
            type: object
      summary: Создание нового пользователя (только для администраторов)
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"